Refactored part of nav component into desktop nav
