+++ src/components/common/navigation/MobileNavigation.tsx
@@ -1,0 +1,4 @@
+
+export function MobileNavigation() {
+    
+}
